Refactor transitional shards into coherent modules
- Collapse TabManager_0..7 into src/tabManager/{core,viewModes,echoLayouts,echoRender}.js
  and merge duplicate layout chunk3/4 bodies that previously overwrote each other
- Extract main_init_4 echo interactions into interactions/EchoDocumentInteractions.js
  with init/destroy lifecycle; main_init_4.js now boots dispatcher only
- Add appContext.js for shared manager references (tabManager, workspaceSession, etc.)
- Merge Cabinet3D_0..1 into cabinet3d/methods.js with proper imports
- Merge styles_1..15 into styles_{base,views,echo,ui}.css; fix unclosed CSS block
- Update AGENTS.md and TabManager/Cabinet3D .d.ts siblings
