+++ index.html
@@ -10,21 +10,10 @@
       href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&family=JetBrains+Mono:wght@400;700&display=swap"
       rel="stylesheet"
     />
-    <link rel="stylesheet" href="/src/styles_1.css" />
-    <link rel="stylesheet" href="/src/styles_2.css" />
-    <link rel="stylesheet" href="/src/styles_3.css" />
-    <link rel="stylesheet" href="/src/styles_4.css" />
-    <link rel="stylesheet" href="/src/styles_5.css" />
-    <link rel="stylesheet" href="/src/styles_6.css" />
-    <link rel="stylesheet" href="/src/styles_7.css" />
-    <link rel="stylesheet" href="/src/styles_8.css" />
-    <link rel="stylesheet" href="/src/styles_9.css" />
-    <link rel="stylesheet" href="/src/styles_10.css" />
-    <link rel="stylesheet" href="/src/styles_11.css" />
-    <link rel="stylesheet" href="/src/styles_12.css" />
-    <link rel="stylesheet" href="/src/styles_13.css" />
-    <link rel="stylesheet" href="/src/styles_14.css" />
-    <link rel="stylesheet" href="/src/styles_15.css" />
+    <link rel="stylesheet" href="/src/styles_base.css" />
+    <link rel="stylesheet" href="/src/styles_views.css" />
+    <link rel="stylesheet" href="/src/styles_echo.css" />
+    <link rel="stylesheet" href="/src/styles_ui.css" />
   </head>
   <body>
     <div id="app">
