--- src/styles_15.css
@@ -1,84 +1,0 @@
-/* Keyboard cheatsheet overlay (generated by CheatsheetOverlay.js). */
-#keyboard-cheatsheet {
-  position: fixed;
-  inset: 0;
-  z-index: 9999;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: rgba(2, 6, 18, 0.72);
-  backdrop-filter: blur(6px);
-  -webkit-backdrop-filter: blur(6px);
-  font-family: "SF Mono", "JetBrains Mono", ui-monospace, monospace;
-  animation: cheatsheet-fade 0.15s ease-out;
-}
-
-@keyframes cheatsheet-fade {
-  from { opacity: 0; }
-  to { opacity: 1; }
-}
-
-#keyboard-cheatsheet .cheatsheet-panel {
-  max-width: min(1100px, 92vw);
-  max-height: 86vh;
-  overflow-y: auto;
-  padding: 28px 32px;
-  border-radius: 14px;
-  border: 1px solid rgba(90, 200, 250, 0.35);
-  background: linear-gradient(160deg, rgba(10, 18, 34, 0.96), rgba(6, 10, 22, 0.96));
-  box-shadow: 0 24px 80px rgba(0, 0, 0, 0.6), 0 0 0 1px rgba(90, 200, 250, 0.08);
-  color: #e6f1ff;
-}
-
-#keyboard-cheatsheet h2 {
-  margin: 0 0 4px;
-  font-size: 20px;
-  letter-spacing: 0.04em;
-  color: #7fdbff;
-}
-
-#keyboard-cheatsheet .cheatsheet-hint {
-  margin: 0 0 20px;
-  font-size: 12px;
-  opacity: 0.6;
-}
-
-#keyboard-cheatsheet .cheatsheet-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
-  gap: 18px 28px;
-}
-
-#keyboard-cheatsheet section h3 {
-  margin: 0 0 8px;
-  font-size: 12px;
-  text-transform: uppercase;
-  letter-spacing: 0.08em;
-  color: #9ad;
-  border-bottom: 1px solid rgba(90, 200, 250, 0.18);
-  padding-bottom: 4px;
-}
-
-#keyboard-cheatsheet .cheatsheet-row {
-  display: flex;
-  align-items: baseline;
-  gap: 12px;
-  padding: 3px 0;
-  font-size: 13px;
-}
-
-#keyboard-cheatsheet .cheatsheet-row kbd {
-  flex: 0 0 auto;
-  min-width: 92px;
-  padding: 2px 8px;
-  border-radius: 5px;
-  border: 1px solid rgba(90, 200, 250, 0.35);
-  background: rgba(90, 200, 250, 0.1);
-  color: #bfe9ff;
-  font-size: 12px;
-  text-align: center;
-}
-
-#keyboard-cheatsheet .cheatsheet-row span {
-  opacity: 0.85;
-}
